gui/requirements.txt
paho-mqtt
PyYAML
pyyaml-include
PyQt5
pyqtgraph
PySDL2
pysdl2-dll
opencv-python